added comments and filtering elimination
I've added more comments, which we can use when writing the report.
We could have used bluring (with a gausian, but ideally with a box filter - all weighted to 1) to make the neighboring pixels influence the intensity, and then apply histogram equalization...
but that will make the resulting thresholding even more round.
I liked the idea you had with morphological operations to adjust the blobs... although they still make the results more 'circular'.
Anyway, in order to check for circularity we needed to check if perimeter/sqrt(pi*area) ~ 0.99... but it appears that the perimeter is not well calculated using the SIGBTools... and it might be harder to calculate...
Therefore, we can use the angle of the ellipses - 90 degrees means that a blob is even a circle or a cat/snake-eye-like blob. That should suffice, because the face doesn't have so many cat/snake-eye-like blobs... even a circle is not 100% correct.